refactor: move files to lib
