+++ .gitignore
@@ -68,7 +68,6 @@ typings/
 # package-specific generated files
 packages/gasket-redux/lib
 packages/gasket-assets/react
-packages/gasket-react-intl/lib
 packages/gasket-plugin-docs-graphs/test/fixtures/generated-docs
 packages/gasket-cli/test/fixtures/create/*
 !packages/gasket-cli/test/fixtures/create/.gitkeep
