+++ packages/gasket-react-intl/lib/index.js
