+++ packages/gasket-react-intl/lib/with-locale-file-required.js
