+++ packages/gasket-react-intl/lib/with-messages-provider.js
@@ -1,3 +1,5 @@
+'use client';
+
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import htmlescape from 'htmlescape';
 import { createElement, useState } from 'react';
