+++ packages/gasket-react-intl/package.json
@@ -6,8 +6,7 @@
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "files": [
-    "lib",
-    "types"
+    "lib"
   ],
   "scripts": {
     "lint": "eslint .",
